Sbp.Data: Change key datatype

+++ src/Folio3.DotNet.Sbp/Folio3.DotNet.Sbp.Data/School/Entities/Course.cs
@@ -11,7 +11,7 @@ namespace Folio3.DotNet.Sbp.Data.School.Entities
     {
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
         [Display(Name = "Number")]
-        public int CourseID { get; set; }
+        public long CourseID { get; set; }
 
         [StringLength(50, MinimumLength = 3)]
         public string Title { get; set; }
@@ -19,7 +19,7 @@ namespace Folio3.DotNet.Sbp.Data.School.Entities
         [Range(0, 5)]
         public int Credits { get; set; }
 
-        public int DepartmentID { get; set; }
+        public long DepartmentID { get; set; }
 
         public Department Department { get; set; }
         public ICollection<Enrollment> Enrollments { get; set; }
